chore: Move delete signing key endpoint to endpoints section of app
